requirements.txt
flask
flask-cors
gunicorn
openai
supabase
requests
python-dotenv
pandas
numpy
